+++ gsa/src/web/pages/scanconfigs/component.js
@@ -500,6 +500,7 @@ class ScanConfigComponent extends React.Component {
         {editConfigFamilyDialogVisible && (
           <EditConfigFamilyDialog
             config={config}
+            configNameLabel={_('Config')}
             config_name={config_name}
             family_name={family_name}
             id={id}
@@ -514,6 +515,7 @@ class ScanConfigComponent extends React.Component {
         {editNvtDetailsDialogVisible && (
           <EditNvtDetailsDialog
             config={config}
+            configNameLabel={_('Config')}
             config_name={config_name}
             family_name={family_name}
             manual_timeout={manual_timeout}