+++ gsa/src/web/pages/scanconfigs/editconfigfamilydialog.js
@@ -197,6 +197,7 @@ class EditDialogComponent extends React.Component {
       config_name,
       family_name,
       id,
+      configNameLabel,
       title,
       onClose,
       onEditNvtDetailsClick,
@@ -227,7 +228,7 @@ class EditDialogComponent extends React.Component {
             <SimpleTable>
               <TableBody>
                 <TableRow>
-                  <TableData>{_('Config')}</TableData>
+                  <TableData>{configNameLabel}</TableData>
                   <TableData>{config_name}</TableData>
                 </TableRow>
                 <TableRow>
@@ -307,6 +308,7 @@ class EditDialogComponent extends React.Component {
 
 EditDialogComponent.propTypes = {
   config: PropTypes.model.isRequired,
+  configNameLabel: PropTypes.string.isRequired,
   config_name: PropTypes.string,
   family_name: PropTypes.string,
   id: PropTypes.string,